Rewrite the native binding in N-API; add context safety in the process.
This is an ongoing nightmare!

It took me a while to realize that the original binding was a relic of how early it had been written in the life of Node. It used a lot of `libuv` functions directly for things that have had easier-to-use abstractions for a while now — both in `nan` and in N-API.

It was also a good opportunity to decaffeinate the JavaScript files. Furthermore, we don't need the `HandleMap` stuff, nor the scarily-named `UnsafePersistent`, since we have proper `Map`s in JavaScript now. (I'm not sure we _ever_ needed `HandleMap`, but I digress.)

This isn't over! It works on macOS, but it needs platform-specific adjustments for Windows and Linux. Both of those will be painful. I'm sure I can also improve upon the macOS file-watching somehow.