Convert the specs to JavaScript.
This comes with the painful revelation that one spec has been spuriously passing for years, and the functionality it claims to support doesn't seem to work (at least on macOS). That spec is skipped until that bug can be fixed.

This required a more modern version of Jasmine that could grok `async`/`await` syntax. The magical `jasmine-tagged` stuff is currently being done manually instead.